backend/requirements.txt
flask
flask-cors
pymupdf
pytesseract
pillow
werkzeug
gunicorn
